Explicitly install a 3 node cluster.
Prior to this change, our install config listed a number of workers.
In practice though, we don't support deploying workers at install
time.  We require scripts post-install to set up the
baremetal-operator before workers can be deployed.  This change makes
it so we expliclty install a 3 node cluster, and then scale it out as
a post-install step if dev-scripts was configured to do so.

A side effect of this change is that the installer will automatically
adjust the scheduler configuration to make master nodes schedulable.
For more details on that change, see:
https://github.com/openshift/installer/pull/2004

This also means that we can drop the custom IngressController
manifest.  When the cluster is configured with "mastersSchedulable:
true" in the scheduler config, the master Nodes will have both the
master and worker roles set on them, making them candidates for
running the default Ingress controller, without any changes.

Closes issue #705